Change of strategy
A huge strange of strategy: instead of simulate click events on the UI button, the plugin will directly interact with the JSONRPC server. Now, the user can select whatever webUI he wants. Moreover, the data shown are more accurate.
However, it consumes a lot more network ressources. The updateInteval at the top is meant to change the update interval